+++ pub/consts.js
@@ -5,8 +5,11 @@ module.exports = {
   ],
   formUrlType: 'application/x-www-form-urlencoded',
   jsonldTypes: [
+    // req.accepts uses accepts which does match profile
     'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
-    'application/activity+json'
+    'application/activity+json',
+    // req.is uses type-is which cannot handle profile
+    'application/ld+json'
   ],
   // type-is is not able to match this pattern
   jsonldOutgoingType: 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',